+++ app/src/main/java/com/mightyfour/application/WeightedSumStrategy.java
@@ -4,7 +4,7 @@ import com.mightyfour.domain.ImpactCalculationStrategy;
 import com.mightyfour.domain.Product;
 
 public class WeightedSumStrategy implements ImpactCalculationStrategy{
-    public int calculateImpact(Product product){
-        return 0; //Not implemented yet.
+    public double calculateImpact(Product product){
+        return 0.0; //Not implemented yet.
     }
 }
