+++ app/src/main/java/com/mightyfour/domain/ImpactCalculationStrategy.java
@@ -1,5 +1,5 @@
 package com.mightyfour.domain;
 
 public interface ImpactCalculationStrategy {
-    int calculateImpact(Product product);
+    double calculateImpact(Product product);
 }
